Revert "X2-7602 reworked range datepicker (#280)"
This reverts commit 3a8448da0b7612aa6ab9349eb0d8d42e6cfb3884.
